widget/tests/chrome_context_menus_win.xul

<?xml version="1.0"?>

<!-- ***** BEGIN LICENSE BLOCK *****
   - Version: MPL 1.1/GPL 2.0/LGPL 2.1
   -
   - The contents of this file are subject to the Mozilla Public License Version
   - 1.1 (the "License"); you may not use this file except in compliance with
   - the License. You may obtain a copy of the License at
   - http://www.mozilla.org/MPL/
   -
   - Software distributed under the License is distributed on an "AS IS" basis,
   - WITHOUT WARRANTY OF ANY KIND, either express or implied. See the License
   - for the specific language governing rights and limitations under the
   - License.
   -
   - The Original Code is Native Menus Test code
   -
   - The Initial Developer of the Original Code is
   - Mozilla Foundation.
   - Portions created by the Initial Developer are Copyright (C) 2010
   - the Initial Developer. All Rights Reserved.
   -
   - Contributor(s):
   -
   - Alternatively, the contents of this file may be used under the terms of
   - either the GNU General Public License Version 2 or later (the "GPL"), or
   - the GNU Lesser General Public License Version 2.1 or later (the "LGPL"),
   - in which case the provisions of the GPL or the LGPL are applicable instead
   - of those above. If you wish to allow use of your version of this file only
   - under the terms of either the GPL or the LGPL, and not to allow others to
   - use your version of this file under the terms of the MPL, indicate your
   - decision by deleting the provisions above and replace them with the notice
   - and other provisions required by the GPL or the LGPL. If you do not delete
   - the provisions above, a recipient may use your version of this file under
   - the terms of any one of the MPL, the GPL or the LGPL.
   -
   - ***** END LICENSE BLOCK ***** -->

<?xml-stylesheet href="chrome://global/skin" type="text/css"?>

<window id="ChromeContextMenuTest"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        width="300"
        height="300"
        title="Chrome Context Menu Test w/Plugin Focus">

<script type="application/javascript"
        src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js" />
<script type="application/javascript"
        src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>

<popupset>
  <menupopup id="testmenu" onpopupshown="menuDisplayed()">
    <menuitem label="One"/>
    <menuitem label="Two"/>
    <menuitem label="Three"/>
  </menupopup>
</popupset>

<toolbox>
  <toolbar id="nav-toolbar" style="height:30px" context="testmenu">
  </toolbar>
</toolbox>

<script type="application/javascript"><![CDATA[

function ok(condition, message) {
    window.opener.wrappedJSObject.SimpleTest.ok(condition, message);
}

function onTestsFinished() {
    window.close();
    window.opener.wrappedJSObject.SimpleTest.finish();
}

function openContextMenuFor(element) {
    netscape.security.PrivilegeManager.enablePrivilege('UniversalXPConnect');

    var utils = window.QueryInterface(Components.interfaces.nsIInterfaceRequestor).
                       getInterface(Components.interfaces.nsIDOMWindowUtils);

    var tbX = (window.mozInnerScreenX + 10) * utils.screenPixelsPerCSSPixel;
    var tbY = (window.mozInnerScreenY + 10) * utils.screenPixelsPerCSSPixel;

    // See nsWidnow's SynthesizeNativeMouseEvent & SendInput on msdn
    var MOUSEEVENTF_RIGHTDOWN = 0x0008;
    var MOUSEEVENTF_RIGHTUP = 0x0010;

    utils.sendNativeMouseEvent(tbX, tbY,
                               MOUSEEVENTF_RIGHTDOWN,
                               0, element);
    utils.sendNativeMouseEvent(tbX, tbY,
                               MOUSEEVENTF_RIGHTUP,
                               0, element);
}

var tid = 0;

function onFocus() {
    var _delayedOnLoad = function() {
        var plugin = document.getElementById("plugin");
        var toolbar = document.getElementById("nav-toolbar");

        plugin.focus();

        tid = setTimeout("menuTimeout()", 5000);

        openContextMenuFor(toolbar);
    }
    setTimeout(_delayedOnLoad, 3000);
}

function menuTimeout() {
    ok(false, "Right-click chrome menu did not display with focus on a plugin.");
    onTestsFinished();
}

function menuDisplayed() {
    clearTimeout(tid);
    ok(true, "Right-click chrome menu displayed despite plugin having focus.");
    onTestsFinished();
}

window.opener.wrappedJSObject.SimpleTest.waitForFocus(onFocus, window);


]]></script>

<body xmlns="http://www.w3.org/1999/xhtml">
  <br/>
  <embed id="plugin" type="application/x-test" width="50" height="50"></embed>
</body>

</window>
